+++ pkg/netpol/connlist/internal/ingressanalyzer/ingress_analyzer.go
@@ -28,7 +28,7 @@ import (
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/common"
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/eval"
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/logger"
-	"github.com/np-guard/netpol-analyzer/pkg/netpol/scan"
+	"github.com/np-guard/netpol-analyzer/pkg/netpol/manifests/parser"
 )
 
 const (
@@ -69,7 +69,7 @@ type IngressAnalyzer struct {
 }
 
 // NewIngressAnalyzerWithObjects returns a new IngressAnalyzer with relevant objects
-func NewIngressAnalyzerWithObjects(objects []scan.K8sObject, pe *eval.PolicyEngine, l logger.Logger,
+func NewIngressAnalyzerWithObjects(objects []parser.K8sObject, pe *eval.PolicyEngine, l logger.Logger,
 	muteErrsAndWarns bool) (*IngressAnalyzer, error) {
 	ia := &IngressAnalyzer{
 		servicesToPortsAndPeersMap: make(map[string]map[string]peersAndPorts),
@@ -83,11 +83,11 @@ func NewIngressAnalyzerWithObjects(objects []scan.K8sObject, pe *eval.PolicyEngi
 	var err error
 	for _, obj := range objects {
 		switch obj.Kind {
-		case scan.Service:
+		case parser.Service:
 			err = ia.mapServiceToPeers(obj.Service)
-		case scan.Route:
+		case parser.Route:
 			ia.mapRouteToServices(obj.Route)
-		case scan.Ingress:
+		case parser.Ingress:
 			ia.mapK8sIngressToServices(obj.Ingress)
 		}
 		if err != nil {
@@ -134,7 +134,7 @@ func (ia *IngressAnalyzer) mapServiceToPeers(svc *corev1.Service) error {
 func (ia *IngressAnalyzer) getServiceSelectedPeers(svc *corev1.Service) ([]eval.Peer, error) {
 	svcStr := types.NamespacedName{Name: svc.Name, Namespace: svc.Namespace}.String()
 	if svc.Spec.Selector == nil {
-		ia.logWarning("Ignoring " + scan.Service + whiteSpace + svcStr + colon + missingSelectorWarning)
+		ia.logWarning("Ignoring " + parser.Service + whiteSpace + svcStr + colon + missingSelectorWarning)
 		return nil, nil
 	}
 	svcLabelsSelector, err := convertServiceSelectorToLabelSelector(svc.Spec.Selector)
@@ -159,7 +159,7 @@ func convertServiceSelectorToLabelSelector(svcSelector map[string]string) (label
 // routes analysis
 
 const (
-	allowedTargetKind      = scan.Service
+	allowedTargetKind      = parser.Service
 	routeTargetKindWarning = "ignoring target with unsupported kind"
 	routeBackendsWarning   = "ignoring alternate backend without Service "
 )
@@ -186,14 +186,14 @@ func (ia *IngressAnalyzer) getRouteServices(rt *ocroutev1.Route) []serviceInfo {
 	}
 	// Currently, only 'Service' is allowed as the kind of target that the route is referring to.
 	if rt.Spec.To.Kind != "" && rt.Spec.To.Kind != allowedTargetKind {
-		ia.logWarning(scan.Route + whiteSpace + routeStr + colon + routeTargetKindWarning)
+		ia.logWarning(parser.Route + whiteSpace + routeStr + colon + routeTargetKindWarning)
 	} else {
 		targetServices = append(targetServices, serviceInfo{serviceName: rt.Spec.To.Name, servicePort: routeTargetPort})
 	}
 
 	for _, backend := range rt.Spec.AlternateBackends {
 		if backend.Kind != "" && backend.Kind != allowedTargetKind {
-			ia.logWarning(scan.Route + whiteSpace + routeStr + colon + routeBackendsWarning)
+			ia.logWarning(parser.Route + whiteSpace + routeStr + colon + routeBackendsWarning)
 		} else {
 			targetServices = append(targetServices, serviceInfo{serviceName: backend.Name, servicePort: routeTargetPort})
 		}
@@ -228,7 +228,7 @@ func (ia *IngressAnalyzer) getK8sIngressServices(ing *netv1.Ingress) []serviceIn
 	// if DefaultBackend is provided add its service info to the result
 	if ing.Spec.DefaultBackend != nil {
 		if ing.Spec.DefaultBackend.Service == nil {
-			ia.logWarning(scan.Ingress + whiteSpace + ingressStr + colon + defaultBackendWarning)
+			ia.logWarning(parser.Ingress + whiteSpace + ingressStr + colon + defaultBackendWarning)
 		} else {
 			backendServices = append(backendServices, getServiceInfo(ing.Spec.DefaultBackend.Service))
 		}
@@ -237,7 +237,7 @@ func (ia *IngressAnalyzer) getK8sIngressServices(ing *netv1.Ingress) []serviceIn
 	for _, rule := range ing.Spec.Rules {
 		for _, path := range rule.IngressRuleValue.HTTP.Paths {
 			if path.Backend.Service == nil {
-				ia.logWarning(scan.Ingress + whiteSpace + ingressStr + colon + ruleBackendWarning)
+				ia.logWarning(parser.Ingress + whiteSpace + ingressStr + colon + ruleBackendWarning)
 			} else {
 				backendServices = append(backendServices, getServiceInfo(path.Backend.Service))
 			}
@@ -277,12 +277,12 @@ func (ia *IngressAnalyzer) AllowedIngressConnections() (map[string]*PeerAndIngre
 
 	// get all targeted workload peers and compute allowed conns of each workload peer
 	// 1. from routes
-	routesResult, err := ia.allowedIngressConnectionsByResourcesType(ia.routesToServicesMap, scan.Route)
+	routesResult, err := ia.allowedIngressConnectionsByResourcesType(ia.routesToServicesMap, parser.Route)
 	if err != nil {
 		return nil, err
 	}
 	// 2. from k8s-ingress objects
-	ingressResult, err := ia.allowedIngressConnectionsByResourcesType(ia.k8sIngressToServicesMap, scan.Ingress)
+	ingressResult, err := ia.allowedIngressConnectionsByResourcesType(ia.k8sIngressToServicesMap, parser.Ingress)
 	if err != nil {
 		return nil, err
 	}
