+++ pkg/netpol/manifests/fsscanner/manifests_test.go
@@ -1,4 +1,4 @@
-package manifests
+package fsscanner
 
 import (
 	"fmt"
@@ -9,17 +9,17 @@ import (
 
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/internal/testutils"
 	"github.com/np-guard/netpol-analyzer/pkg/netpol/logger"
-	"github.com/np-guard/netpol-analyzer/pkg/netpol/scan"
+	"github.com/np-guard/netpol-analyzer/pkg/netpol/manifests/parser"
 )
 
 //nolint:gocritic //temporary commented-out code
 func TestBasic(t *testing.T) {
-	dirPath := filepath.Join(testutils.GetTestsDir(), "basic")
+	dirPath := filepath.Join(testutils.GetTestsDirWithDepth(4), "basic")
 	rList, errs := GetResourceInfosFromDirPath([]string{dirPath}, true, false)
 	require.Empty(t, errs, "expecting no errors on basic dir")
 
 	// TODO: move the code below to scan pkg
-	oList, _ := scan.ResourceInfoListToK8sObjectsList(rList, logger.NewDefaultLogger(), false)
+	oList, _ := parser.ResourceInfoListToK8sObjectsList(rList, logger.NewDefaultLogger(), false)
 	// require.Nil(t, err, "err ResourceInfoToK8sObjects")
 	require.Equal(t, len(oList), len(rList), "expecting same length fot input and output lists")
 	fmt.Println("done")
