--- pkg/netpol/scan/podList.yaml
@@ -1,79 +1,0 @@
-apiVersion: v1
-items:
-- apiVersion: v1
-  kind: Pod
-  metadata:
-    creationTimestamp: "2022-08-01T17:05:22Z"
-    generateName: adservice-77d5cd745d-
-    labels:
-      app: adservice
-      pod-template-hash: 77d5cd745d
-    name: adservice-77d5cd745d-t8mx4
-    namespace: default
-    ownerReferences:
-    - apiVersion: apps/v1
-      blockOwnerDeletion: true
-      controller: true
-      kind: ReplicaSet
-      name: adservice-77d5cd745d
-      uid: de20939b-c550-4dea-943d-57140c16d9e4
-    resourceVersion: "1632"
-    uid: 4c85f10a-15b3-4a57-bd57-c3f05240068c
-  spec:
-    containers:
-    - env:
-      - name: PORT
-        value: "9555"
-      - name: DISABLE_STATS
-        value: "1"
-      imagePullPolicy: IfNotPresent
-      livenessProbe:
-        exec:
-          command:
-          - /bin/grpc_health_probe
-          - -addr=:9555
-        failureThreshold: 3
-        initialDelaySeconds: 20
-        periodSeconds: 15
-        successThreshold: 1
-        timeoutSeconds: 1
-      name: server
-      ports:
-      - containerPort: 9555
-        protocol: TCP
-      readinessProbe:
-        exec:
-          command:
-          - /bin/grpc_health_probe
-          - -addr=:9555
-        failureThreshold: 3
-        initialDelaySeconds: 20
-        periodSeconds: 15
-        successThreshold: 1
-        timeoutSeconds: 1
-      resources:
-        limits:
-          cpu: 300m
-          memory: 300Mi
-        requests:
-          cpu: 200m
-          memory: 180Mi
-      terminationMessagePath: /dev/termination-log
-      terminationMessagePolicy: File
-    dnsPolicy: ClusterFirst
-    enableServiceLinks: true
-    nodeName: minikube
-    preemptionPolicy: PreemptLowerPriority
-    priority: 0
-    restartPolicy: Always
-    schedulerName: default-scheduler
-    securityContext: {}
-    serviceAccount: default
-    serviceAccountName: default
-    terminationGracePeriodSeconds: 5
-  status:
-    hostIP: 192.168.49.2
-kind: List
-metadata:
-    resourceVersion: ""
-    selfLink: ""
